agent/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
markdown==3.5.2
pypandoc==1.14
beautifulsoup4==4.12.3
pandas>=2.0.0
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3
Pillow==11.1.0
tiktoken>=0.5.1
matplotlib>=3.7.0
